Change Joi validation, its an object no function. Validation work fine but need remove _id type from schema. Mongoose should add _id.
